serial-proxy/requirements.txt
pyserial
pubnub>=3.7.5
m2x>=3.0.4